[SL] Add another monaco editor using a diff way
