+++ frontend/src/pages/room.tsx
@@ -1,5 +1,6 @@
 import Head from 'next/head'
 import Layout from '@/components/Layout/Layout'
+import OldCodeEditor from '@/components/CodeEditor/OldCodeEditor'
 import CodeEditor from '@/components/CodeEditor/CodeEditor'
 
 export default function CodeRoom() {
